+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 Features
   - Add environment variables module recipe [#435](https://github.com/platanus/potassium/pull/435)
+  - Run [Brakeman](https://brakemanscanner.org/) with reviewdog on CI [#440](https://github.com/platanus/potassium/pull/440)
 
 Fixes
   - Fix CircleCI config [#434](https://github.com/platanus/potassium/pull/434)